Move OrderT, Prepend, NonEmpty to consensus.  Organize consensus module into algebras/interpreters/models.  Move BodyToHeaderValidation to blockchain.
