+++ consensus/src/main/scala/co/topl/consensus/interpreters/EtaCalculation.scala
@@ -17,134 +17,131 @@ import scalacache.caffeine.CaffeineCache
 
 object EtaCalculation {
 
-  object Eval {
+  def make[F[_]: Sync: Logger](
+    fetchSlotData:      TypedIdentifier => F[SlotData],
+    clock:              ClockAlgebra[F],
+    genesisEta:         Eta,
+    blake2b256Resource: UnsafeResource[F, Blake2b256],
+    blake2b512Resource: UnsafeResource[F, Blake2b512]
+  ): F[EtaCalculationAlgebra[F]] =
+    for {
+      implicit0(cache: CaffeineCache[F, Bytes, Eta]) <- CaffeineCache[F, Bytes, Eta]
+      slotsPerEpoch                                  <- clock.slotsPerEpoch
+      impl = new Impl[F](fetchSlotData, clock, genesisEta, slotsPerEpoch, blake2b256Resource, blake2b512Resource)
+    } yield impl
 
-    def make[F[_]: Sync: Logger](
-      fetchSlotData:      TypedIdentifier => F[SlotData],
-      clock:              ClockAlgebra[F],
-      genesisEta:         Eta,
-      blake2b256Resource: UnsafeResource[F, Blake2b256],
-      blake2b512Resource: UnsafeResource[F, Blake2b512]
-    ): F[EtaCalculationAlgebra[F]] =
-      for {
-        implicit0(cache: CaffeineCache[F, Bytes, Eta]) <- CaffeineCache[F, Bytes, Eta]
-        slotsPerEpoch                                  <- clock.slotsPerEpoch
-        impl = new Impl[F](fetchSlotData, clock, genesisEta, slotsPerEpoch, blake2b256Resource, blake2b512Resource)
-      } yield impl
+  private class Impl[F[_]: Sync: Logger](
+    fetchSlotData:      TypedIdentifier => F[SlotData],
+    clock:              ClockAlgebra[F],
+    genesisEta:         Eta,
+    slotsPerEpoch:      Long,
+    blake2b256Resource: UnsafeResource[F, Blake2b256],
+    blake2b512Resource: UnsafeResource[F, Blake2b512]
+  )(implicit cache:     CaffeineCache[F, Bytes, Eta])
+      extends EtaCalculationAlgebra[F] {
 
-    private class Impl[F[_]: Sync: Logger](
-      fetchSlotData:      TypedIdentifier => F[SlotData],
-      clock:              ClockAlgebra[F],
-      genesisEta:         Eta,
-      slotsPerEpoch:      Long,
-      blake2b256Resource: UnsafeResource[F, Blake2b256],
-      blake2b512Resource: UnsafeResource[F, Blake2b512]
-    )(implicit cache:     CaffeineCache[F, Bytes, Eta])
-        extends EtaCalculationAlgebra[F] {
+    private val twoThirdsLength = slotsPerEpoch * 2 / 3
 
-      private val twoThirdsLength = slotsPerEpoch * 2 / 3
-
-      override def etaToBe(parentSlotId: SlotId, childSlot: Slot): F[Eta] =
-        if (childSlot < slotsPerEpoch) genesisEta.pure[F]
-        else
-          (
-            clock.epochOf(parentSlotId.slot),
-            clock.epochOf(childSlot),
-            fetchSlotData(parentSlotId.blockId)
-          ).tupled
-            .flatMap {
-              case (parentEpoch, childEpoch, parentSlotData) if parentEpoch === childEpoch =>
-                parentSlotData.eta.pure[F]
-              // TODO: If childEpoch - parentEpoch > 1, destroy the node
-              // OR: childSlot - parentSlot > slotsPerEpoch
-              case (_, _, parentSlotData) =>
-                cache.cachingF(parentSlotId.blockId.allBytes)(ttl = None)(
-                  Sync[F].defer(
-                    locateTwoThirdsBest(parentSlotData)
-                      .flatMap(calculate)
-                      .flatTap(nextEta =>
-                        Logger[F]
-                          .info(show"Caching child epoch's eta for parent id=${parentSlotId.blockId} nextEta=$nextEta")
-                      )
-                  )
+    override def etaToBe(parentSlotId: SlotId, childSlot: Slot): F[Eta] =
+      if (childSlot < slotsPerEpoch) genesisEta.pure[F]
+      else
+        (
+          clock.epochOf(parentSlotId.slot),
+          clock.epochOf(childSlot),
+          fetchSlotData(parentSlotId.blockId)
+        ).tupled
+          .flatMap {
+            case (parentEpoch, childEpoch, parentSlotData) if parentEpoch === childEpoch =>
+              parentSlotData.eta.pure[F]
+            // TODO: If childEpoch - parentEpoch > 1, destroy the node
+            // OR: childSlot - parentSlot > slotsPerEpoch
+            case (_, _, parentSlotData) =>
+              cache.cachingF(parentSlotId.blockId.allBytes)(ttl = None)(
+                Sync[F].defer(
+                  locateTwoThirdsBest(parentSlotData)
+                    .flatMap(calculate)
+                    .flatTap(nextEta =>
+                      Logger[F]
+                        .info(show"Caching child epoch's eta for parent id=${parentSlotId.blockId} nextEta=$nextEta")
+                    )
                 )
-            }
+              )
+          }
 
-      /**
-       * Given some header near the end of an epoch, traverse the chain (toward genesis) until reaching a block
-       * that is inside of the 2/3 window of the epoch
-       */
-      private def locateTwoThirdsBest(from: SlotData): F[SlotData] =
-        if (isWithinTwoThirds(from)) from.pure[F]
-        else
-          from
-            .iterateUntilM(data => fetchSlotData(data.parentSlotId.blockId))(isWithinTwoThirds)
-            .flatTap(twoThirdsBest =>
-              Logger[F].info(show"Located twoThirdsBest=${twoThirdsBest.slotId.blockId} from=${from.slotId.blockId}")
-            )
+    /**
+     * Given some header near the end of an epoch, traverse the chain (toward genesis) until reaching a block
+     * that is inside of the 2/3 window of the epoch
+     */
+    private def locateTwoThirdsBest(from: SlotData): F[SlotData] =
+      if (isWithinTwoThirds(from)) from.pure[F]
+      else
+        from
+          .iterateUntilM(data => fetchSlotData(data.parentSlotId.blockId))(isWithinTwoThirds)
+          .flatTap(twoThirdsBest =>
+            Logger[F].info(show"Located twoThirdsBest=${twoThirdsBest.slotId.blockId} from=${from.slotId.blockId}")
+          )
 
-      private def isWithinTwoThirds(from: SlotData): Boolean =
-        from.slotId.slot % slotsPerEpoch <= twoThirdsLength
+    private def isWithinTwoThirds(from: SlotData): Boolean =
+      from.slotId.slot % slotsPerEpoch <= twoThirdsLength
 
-      /**
-       * Compute the Eta value for the epoch containing the given header
-       * @param twoThirdsBest The latest block header in some tine, but within the first 2/3 of the epoch
-       */
-      private def calculate(twoThirdsBest: SlotData): F[Eta] =
-        cache.cachingF(twoThirdsBest.slotId.blockId.allBytes)(ttl = None)(
-          Sync[F].defer(
-            for {
-              epoch      <- clock.epochOf(twoThirdsBest.slotId.slot)
-              epochRange <- clock.epochRange(epoch)
-              epochData <- NonEmptyChain(twoThirdsBest).iterateUntilM(items =>
-                fetchSlotData(items.head.parentSlotId.blockId).map(items.prepend)
-              )(items => items.head.parentSlotId.slot < epochRange.start)
-              rhoValues = epochData.map(_.rho)
-              nextEta <- calculate(previousEta = twoThirdsBest.eta, epoch + 1, rhoValues)
-            } yield nextEta
+    /**
+     * Compute the Eta value for the epoch containing the given header
+     * @param twoThirdsBest The latest block header in some tine, but within the first 2/3 of the epoch
+     */
+    private def calculate(twoThirdsBest: SlotData): F[Eta] =
+      cache.cachingF(twoThirdsBest.slotId.blockId.allBytes)(ttl = None)(
+        Sync[F].defer(
+          for {
+            epoch      <- clock.epochOf(twoThirdsBest.slotId.slot)
+            epochRange <- clock.epochRange(epoch)
+            epochData <- NonEmptyChain(twoThirdsBest).iterateUntilM(items =>
+              fetchSlotData(items.head.parentSlotId.blockId).map(items.prepend)
+            )(items => items.head.parentSlotId.slot < epochRange.start)
+            rhoValues = epochData.map(_.rho)
+            nextEta <- calculate(previousEta = twoThirdsBest.eta, epoch + 1, rhoValues)
+          } yield nextEta
+        )
+      )
+
+    /**
+     * Calculate a new Eta value once all the necessary pre-requisites have been gathered
+     */
+    private def calculate(
+      previousEta: Eta,
+      epoch:       Epoch,
+      rhoValues:   NonEmptyChain[Rho]
+    ): F[Eta] =
+      (Logger[F].info(
+        show"Calculating new eta.  previousEta=$previousEta epoch=$epoch rhoValues=[${rhoValues.length}]{${rhoValues.head}..${rhoValues.last}}"
+      ) >>
+        blake2b512Resource
+          .use(implicit blake2b512 => rhoValues.map(_.sizedBytes.data).map(Ed25519VRF.rhoToRhoNonceHash).pure[F])
+          .map(_.map(nonceHashBytes => RhoNonceHash(Sized.strictUnsafe(nonceHashBytes))))
+          .flatMap(calculateFromNonceHashValues(previousEta, epoch, _)))
+        .flatTap(nextEta =>
+          Logger[F].info(
+            show"Finished calculating new eta.  previousEta=$previousEta epoch=$epoch rhoValues=[${rhoValues.length}]{${rhoValues.head}..${rhoValues.last}} nextEta=$nextEta"
           )
         )
 
-      /**
-       * Calculate a new Eta value once all the necessary pre-requisites have been gathered
-       */
-      private def calculate(
-        previousEta: Eta,
-        epoch:       Epoch,
-        rhoValues:   NonEmptyChain[Rho]
-      ): F[Eta] =
-        (Logger[F].info(
-          show"Calculating new eta.  previousEta=$previousEta epoch=$epoch rhoValues=[${rhoValues.length}]{${rhoValues.head}..${rhoValues.last}}"
-        ) >>
-          blake2b512Resource
-            .use(implicit blake2b512 => rhoValues.map(_.sizedBytes.data).map(Ed25519VRF.rhoToRhoNonceHash).pure[F])
-            .map(_.map(nonceHashBytes => RhoNonceHash(Sized.strictUnsafe(nonceHashBytes))))
-            .flatMap(calculateFromNonceHashValues(previousEta, epoch, _)))
-          .flatTap(nextEta =>
-            Logger[F].info(
-              show"Finished calculating new eta.  previousEta=$previousEta epoch=$epoch rhoValues=[${rhoValues.length}]{${rhoValues.head}..${rhoValues.last}} nextEta=$nextEta"
-            )
-          )
-
-      /**
-       * Calculate a new Eta value once all the necessary pre-requisites have been gathered
-       */
-      private def calculateFromNonceHashValues(
-        previousEta:        Eta,
-        epoch:              Epoch,
-        rhoNonceHashValues: NonEmptyChain[RhoNonceHash]
-      ): F[Eta] =
-        blake2b256Resource
-          .use(
-            _.hash(
-              Bytes
-                .concat(EtaCalculationArgs(previousEta, epoch, rhoNonceHashValues.toIterable).digestMessages)
-            ).pure[F]
-          )
-          .map(Sized.strictUnsafe(_): Eta)
-    }
-
+    /**
+     * Calculate a new Eta value once all the necessary pre-requisites have been gathered
+     */
+    private def calculateFromNonceHashValues(
+      previousEta:        Eta,
+      epoch:              Epoch,
+      rhoNonceHashValues: NonEmptyChain[RhoNonceHash]
+    ): F[Eta] =
+      blake2b256Resource
+        .use(
+          _.hash(
+            Bytes
+              .concat(EtaCalculationArgs(previousEta, epoch, rhoNonceHashValues.toIterable).digestMessages)
+          ).pure[F]
+        )
+        .map(Sized.strictUnsafe(_): Eta)
   }
+
 }
 
 private case class EtaCalculationArgs(previousEta: Eta, epoch: Epoch, rhoNonceHashValues: Iterable[RhoNonceHash]) {
