+++ consensus/src/main/scala/co/topl/consensus/interpreters/LeaderElectionValidation.scala
@@ -22,67 +22,64 @@ object LeaderElectionValidation {
   implicit val signableVrfArgument: Signable[VrfArgument] =
     arg => arg.eta.data ++ Bytes(BigInt(arg.slot).toByteArray)
 
-  object Eval {
+  /**
+   * Normalization constant for test nonce hash evaluation based on 512 byte hash function output
+   */
+  private val NormalizationConstant: BigInt = BigInt(2).pow(512)
 
-    /**
-     * Normalization constant for test nonce hash evaluation based on 512 byte hash function output
-     */
-    private val NormalizationConstant: BigInt = BigInt(2).pow(512)
+  def make[F[_]: Sync](
+    config:             VrfConfig,
+    blake2b512Resource: UnsafeResource[F, Blake2b512],
+    exp:                Exp[F],
+    log1p:              Log1p[F]
+  ): LeaderElectionValidationAlgebra[F] =
+    new LeaderElectionValidationAlgebra[F] {
 
-    def make[F[_]: Sync](
-      config:             VrfConfig,
-      blake2b512Resource: UnsafeResource[F, Blake2b512],
-      exp:                Exp[F],
-      log1p:              Log1p[F]
-    ): LeaderElectionValidationAlgebra[F] =
-      new LeaderElectionValidationAlgebra[F] {
+      def getThreshold(relativeStake: Ratio, slotDiff: Long): F[Ratio] = {
 
-        def getThreshold(relativeStake: Ratio, slotDiff: Long): F[Ratio] = {
+        val difficultyCurve: Ratio =
+          if (slotDiff > config.lddCutoff) config.baselineDifficulty
+          else Ratio(BigInt(slotDiff), BigInt(config.lddCutoff)) * config.amplitude
 
-          val difficultyCurve: Ratio =
-            if (slotDiff > config.lddCutoff) config.baselineDifficulty
-            else Ratio(BigInt(slotDiff), BigInt(config.lddCutoff)) * config.amplitude
-
-          if (difficultyCurve == Ratio.One) {
-            Ratio.One.pure[F]
-          } else
-            for {
-              coefficient <- log1p.evaluate(Ratio.NegativeOne * difficultyCurve)
-              result      <- exp.evaluate(coefficient * relativeStake)
-            } yield Ratio.One - result
-        }
-
-        /**
-         * Determines if the given proof meets the threshold to be elected slot leader
-         * @param threshold the threshold to reach
-         * @param rho the randomness
-         * @return true if elected slot leader and false otherwise
-         */
-        def isSlotLeaderForThreshold(threshold: Ratio)(rho: Rho): F[Boolean] =
-          blake2b512Resource.use(implicit blake2b512 =>
-            Sync[F].delay {
-              val testRhoHashBytes = Ed25519VRF.rhoToRhoTestHash(rho.sizedBytes.data).toArray
-              val test = Ratio(BigInt(Array(0x00.toByte) ++ testRhoHashBytes), NormalizationConstant, BigInt(1))
-              (threshold > test)
-            }
-          )
+        if (difficultyCurve == Ratio.One) {
+          Ratio.One.pure[F]
+        } else
+          for {
+            coefficient <- log1p.evaluate(Ratio.NegativeOne * difficultyCurve)
+            result      <- exp.evaluate(coefficient * relativeStake)
+          } yield Ratio.One - result
       }
 
-    def makeCached[F[_]: Sync](alg: LeaderElectionValidationAlgebra[F]): F[LeaderElectionValidationAlgebra[F]] =
-      CaffeineCache[F, (Ratio, Slot), Ratio].map(cache =>
-        new LeaderElectionValidationAlgebra[F] {
+      /**
+       * Determines if the given proof meets the threshold to be elected slot leader
+       * @param threshold the threshold to reach
+       * @param rho the randomness
+       * @return true if elected slot leader and false otherwise
+       */
+      def isSlotLeaderForThreshold(threshold: Ratio)(rho: Rho): F[Boolean] =
+        blake2b512Resource.use(implicit blake2b512 =>
+          Sync[F].delay {
+            val testRhoHashBytes = Ed25519VRF.rhoToRhoTestHash(rho.sizedBytes.data).toArray
+            val test = Ratio(BigInt(Array(0x00.toByte) ++ testRhoHashBytes), NormalizationConstant, BigInt(1))
+            (threshold > test)
+          }
+        )
+    }
 
-          override def getThreshold(relativeStake: Ratio, slotDiff: Slot): F[Ratio] =
-            cache.cachingF((relativeStake, slotDiff))(ttl = None)(
-              Sync[F].defer(
-                alg.getThreshold(relativeStake, slotDiff)
-              )
+  def makeCached[F[_]: Sync](alg: LeaderElectionValidationAlgebra[F]): F[LeaderElectionValidationAlgebra[F]] =
+    CaffeineCache[F, (Ratio, Slot), Ratio].map(cache =>
+      new LeaderElectionValidationAlgebra[F] {
+
+        override def getThreshold(relativeStake: Ratio, slotDiff: Slot): F[Ratio] =
+          cache.cachingF((relativeStake, slotDiff))(ttl = None)(
+            Sync[F].defer(
+              alg.getThreshold(relativeStake, slotDiff)
             )
+          )
 
-          override def isSlotLeaderForThreshold(threshold: Ratio)(rho: Rho): F[Boolean] =
-            alg.isSlotLeaderForThreshold(threshold)(rho)
-        }
-      )
+        override def isSlotLeaderForThreshold(threshold: Ratio)(rho: Rho): F[Boolean] =
+          alg.isSlotLeaderForThreshold(threshold)(rho)
+      }
+    )
 
-  }
 }
