+++ consensus/src/main/scala/co/topl/consensus/package.scala
@@ -1,6 +1,5 @@
 package co.topl
 
-import cats.Order
 import co.topl.models._
 import co.topl.codecs.bytes.typeclasses.implicits._
 import co.topl.crypto.signing.Ed25519VRF
@@ -8,12 +7,6 @@ import co.topl.models.utility.HasLength.instances.bytesLength
 import co.topl.models.utility.Sized
 
 package object consensus {
-
-  implicit class OrderSupport[T](order: Order[T]) {
-
-    def tiebreakWith(other: Order[T]): Order[T] =
-      Order.whenEqual(order, other)
-  }
 
   implicit class BlockHeaderOps(blockHeader: BlockHeader) {
