+++ consensus/src/test/scala/co/topl/consensus/interpreters/ChainSelectionSpec.scala
@@ -1,11 +1,11 @@
-package co.topl.consensus
+package co.topl.consensus.interpreters
 
 import cats.effect.IO
 import cats.effect.unsafe.implicits.global
 import cats.implicits._
 import co.topl.algebras.UnsafeResource
-import co.topl.crypto.signing.Ed25519VRF
 import co.topl.crypto.hash.Blake2b512
+import co.topl.crypto.signing.Ed25519VRF
 import co.topl.models.ModelGenerators._
 import co.topl.models._
 import co.topl.models.utility.HasLength.instances._
@@ -46,7 +46,7 @@ class ChainSelectionSpec
 
     val orderT =
       ChainSelection
-        .orderT[F](mockFunction[TypedIdentifier, F[SlotData]], blake2b512Resource, kLookback = 1, sWindow = 1)
+        .make[F](mockFunction[TypedIdentifier, F[SlotData]], blake2b512Resource, kLookback = 1, sWindow = 1)
 
     orderT.compare(slotData, slotData).unsafeRunSync() shouldBe 0
   }
@@ -78,7 +78,7 @@ class ChainSelectionSpec
       .anyNumberOfTimes()
       .onCall((id: TypedIdentifier) => allBlocks(id).pure[F])
 
-    val orderT = ChainSelection.orderT[F](fetchSlotData, blake2b512Resource, kLookback = 100, sWindow = 1)
+    val orderT = ChainSelection.make[F](fetchSlotData, blake2b512Resource, kLookback = 100, sWindow = 1)
 
     orderT.compare(xSegment.last, ySegment.last).unsafeRunSync() should be > 0
   }
@@ -116,7 +116,7 @@ class ChainSelectionSpec
       .anyNumberOfTimes()
       .onCall((id: TypedIdentifier) => allBlocks(id).pure[F])
 
-    val orderT = ChainSelection.orderT[F](fetchSlotData, blake2b512Resource, kLookback = 100, sWindow = 1)
+    val orderT = ChainSelection.make[F](fetchSlotData, blake2b512Resource, kLookback = 100, sWindow = 1)
 
     orderT.compare(xSegment.last, ySegment.last).unsafeRunSync() should be > 0
   }
@@ -175,7 +175,7 @@ class ChainSelectionSpec
       .anyNumberOfTimes()
       .onCall((id: TypedIdentifier) => allBlocks(id).pure[F])
 
-    val orderT = ChainSelection.orderT[F](fetchSlotData, blake2b512Resource, kLookback = 100, sWindow = 1)
+    val orderT = ChainSelection.make[F](fetchSlotData, blake2b512Resource, kLookback = 100, sWindow = 1)
 
     orderT.compare(xSegment.last, ySegment.last).unsafeRunSync() should be > 0
   }
@@ -208,7 +208,7 @@ class ChainSelectionSpec
       .anyNumberOfTimes()
       .onCall((id: TypedIdentifier) => allBlocks(id).pure[F])
 
-    val orderT = ChainSelection.orderT[F](fetchSlotData, blake2b512Resource, kLookback = 10, sWindow = 20)
+    val orderT = ChainSelection.make[F](fetchSlotData, blake2b512Resource, kLookback = 10, sWindow = 20)
 
     orderT.compare(xSegment.last, ySegment.last).unsafeRunSync() should be > 0
   }
